fix: remove legacy UI and improve tauri onboarding wizard
