+++ src/e2e/BUILD.bazel
@@ -26,7 +26,7 @@ _NEXT_UI_E2E_SPECS = [
     "//src/ui/next:src/e2e/link-in-bio.spec.ts",
     "//src/ui/next:src/e2e/milestone-loop.spec.ts",
     "//src/ui/next:src/e2e/milestones.spec.ts",
-    "//src/ui/next:src/e2e/onboarding.spec.ts",
+
     "//src/ui/next:src/e2e/review-campaign.spec.ts",
     "//src/ui/next:src/e2e/staff-mesh.spec.ts",
     "//src/ui/next:src/e2e/store-wrapped.spec.ts",
