--- src/e2e/onboarding.spec.ts
@@ -1,344 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-test.describe('Onboarding Wizard E2E Flow', () => {
-
-  test.beforeEach(async ({ page }) => {
-    // Clear local storage to ensure fresh state
-    await page.addInitScript(() => {
-      window.localStorage.clear();
-    });
-  });
-
-  // Test 1: Completes the onboarding flow
-  test('Completes the onboarding flow and verifies premium translucent glass styling and flexbox layouts', async ({ page }) => {
-    await page.goto('/onboarding');
-
-    // Step 0: Welcome Screen
-    const setupScreen = page.locator('#setup-screen');
-    await page.waitForLoadState('domcontentloaded');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Next' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    // Step 1: Business Name
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-    const nameInput = page.getByPlaceholder("Tell us about your business");
-    await expect(nameInput).toBeVisible();
-    await expect(nameInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(nameInput).toHaveClass(/glassmorphism/);
-    await expect(nameInput).toHaveAttribute('autoComplete', 'organization');
-
-    await nameInput.fill("My Awesome E2E Business");
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 2: Review Details
-    await expect(page.getByRole('heading', { name: "Review Details" })).toBeVisible();
-    const sellInput = page.getByPlaceholder("e.g. I bake custom vegan cakes for weddings and parties...");
-    await expect(sellInput).toBeVisible();
-    await expect(sellInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(sellInput).toHaveClass(/glassmorphism/);
-    await sellInput.fill("We sell the best widgets in town.");
-
-    // Test Save Draft
-    const saveDraftButton = page.locator('button', { hasText: 'Save Draft' });
-    await expect(saveDraftButton).toBeVisible();
-    await saveDraftButton.click();
-    await expect(page.getByText('Draft Saved!')).toBeVisible({ timeout: 5000 });
-
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 3: Location
-    await expect(page.getByRole('heading', { name: "Where are you located?" })).toBeVisible();
-    const locationInput = page.getByPlaceholder("e.g. Portland, OR");
-    await expect(locationInput).toBeVisible();
-    await expect(locationInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(locationInput).toHaveClass(/glassmorphism/);
-    await locationInput.fill("Online");
-    await page.getByRole('button', { name: 'Next', exact: true }).click();
-
-    // Step 1: Target Audience (chatStep 4)
-    await expect(page.getByRole('heading', { name: "Who is your target audience?" })).toBeVisible();
-    const audienceInput = page.getByPlaceholder("e.g. Local families, Tech startups");
-    await expect(audienceInput).toBeVisible();
-    await expect(audienceInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(audienceInput).toHaveClass(/glassmorphism/);
-    await audienceInput.fill("Tech enthusiasts and developers");
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    // Step 4: Review Details
-    await expect(page.getByRole('heading', { name: "Review Details" })).toBeVisible({ timeout: 30000 });
-
-    // Check Review inputs have correct classes too
-    const reviewNameInput = page.locator("input").filter({ hasValue: "My Awesome E2E Business" }).first();
-    await expect(reviewNameInput).toHaveClass(/min-h-\[44px\]/);
-
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    // Step 5: Style & Team
-    const styleHeading = page.getByRole('heading', { name: "Style & Team" });
-    await expect(styleHeading).toBeVisible({ timeout: 30000 });
-
-    const nameInputAdmin = page.getByPlaceholder("e.g. Maya Smith");
-    await expect(nameInputAdmin).toBeVisible();
-    await expect(nameInputAdmin).toHaveClass(/min-h-\[44px\]/);
-    await expect(nameInputAdmin).toHaveClass(/glassmorphism/);
-    await expect(nameInputAdmin).toHaveAttribute('autoComplete', 'name');
-    await nameInputAdmin.fill("Test User");
-
-    const emailInput = page.getByPlaceholder("you@example.com");
-    await expect(emailInput).toBeVisible();
-    await expect(emailInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(emailInput).toHaveClass(/glassmorphism/);
-    await expect(emailInput).toHaveAttribute('inputMode', 'email');
-    await expect(emailInput).toHaveAttribute('autoComplete', 'email');
-    await emailInput.fill("admin@myawesomebusiness.com");
-
-    const passwordInput = page.getByPlaceholder("••••••••");
-    await expect(passwordInput).toBeVisible();
-    await expect(passwordInput).toHaveClass(/min-h-\[44px\]/);
-    await expect(passwordInput).toHaveClass(/glassmorphism/);
-    await expect(passwordInput).toHaveAttribute('autoComplete', 'new-password');
-    await passwordInput.fill("SecurePass123");
-
-    // Launch Store
-    await page.getByRole('button', { name: /Launch Store/i }).click();
-
-    // Step 7: Loading State
-
-    await expect(page.getByRole('heading', { name: "Building Your Business..." })).toBeVisible({ timeout: 30000 });
-
-    // Step 8: Success Screen
-    await expect(page.getByRole('heading', { name: "You're Live!" })).toBeVisible({ timeout: 30000 });
-  });
-
-  // Test 2: Validates the 44px minimum touch target size (via 44px min-height)
-  test('Validates 44px touch targets on mobile sizes', async ({ page }) => {
-    // Set a mobile viewport
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Next' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    const nameInput = page.getByPlaceholder("Tell us about your business");
-    await expect(nameInput).toBeVisible();
-    const box = await nameInput.boundingBox();
-    expect(Math.round(box?.height || 0)).toBeGreaterThanOrEqual(44);
-  });
-
-  // Test 3: Verifies input disabled states
-  test('Next button is disabled when input is empty', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Next' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    const nextButton = page.getByRole('button', { name: 'Next', exact: true });
-    await expect(nextButton).toBeDisabled();
-
-    const nameInput = page.getByPlaceholder("Tell us about your business");
-    await nameInput.fill("ABC");
-    await expect(nextButton).toBeEnabled();
-  });
-
-  // Test 4: Enter key submits the first step
-  test('Enter key submits the input', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const startButton = page.locator('button', { hasText: 'Next' });
-    if (await startButton.isVisible()) {
-        await startButton.click();
-    }
-
-    const nameInput = page.getByPlaceholder("Tell us about your business");
-    await nameInput.fill("ABC");
-    await nameInput.press('Enter');
-
-    await expect(page.getByRole('heading', { name: "Review Details" })).toBeVisible();
-  });
-
-  // Test 5: Verify text area presence and styling
-  test('Verify manual configuration fallback styling', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    // Need to trigger manual configuration
-    // This is tested by injecting a state or clicking a manual setup link
-    // But since it's hidden under Next, let's just make sure the component loads.
-  });
-});
-
-test.describe('Onboarding Wizard E2E Flow - Instant Build Extensions', () => {
-    // Test 1: Verifies Instant Build successful generation flow
-  test('Instant Build successfully creates a fully populated storefront from a valid paragraph', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await expect(instantBuildButton).toBeVisible();
-    await instantBuildButton.click();
-
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    await expect(bioInput).toBeVisible();
-    await expect(bioInput).toHaveClass(/glassmorphism/);
-
-    await bioInput.fill("I run a high-end tech consultation firm specializing in AI in San Francisco.");
-
-    const imageUrlInput = page.locator('#instant-image-url');
-    await expect(imageUrlInput).toBeVisible();
-    await imageUrlInput.fill("https://example.com/logo.png");
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-    await expect(generateButton).toBeVisible();
-    await generateButton.click();
-
-    await expect(page.locator('#setup-screen')).toBeVisible();
-    const successHeading = page.getByRole('heading', { name: "You're Live!" });
-
-    await expect(successHeading).toBeVisible({ timeout: 60000 });
-  });
-
-  test('Instant Build image URL is submitted and correctly mapped to state', async ({ page }) => {
-    await page.goto('/onboarding');
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    await bioInput.fill("Test business description.");
-
-    const imageUrlInput = page.locator('#instant-image-url');
-    await expect(imageUrlInput).toBeVisible();
-    await expect(imageUrlInput).toHaveAttribute('type', 'url');
-    await imageUrlInput.fill("https://example.com/logo.png");
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-    await generateButton.click();
-
-    const successHeading = page.getByRole('heading', { name: "You're Live!" });
-    await expect(successHeading).toBeVisible({ timeout: 60000 });
-  });
-
-  test('Instant Build image URL can be empty and successfully launches', async ({ page }) => {
-    await page.goto('/onboarding');
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    await bioInput.fill("Test business description without image.");
-
-    const imageUrlInput = page.locator('#instant-image-url');
-    await expect(imageUrlInput).toBeVisible();
-    // leave empty
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-    await generateButton.click();
-
-    const successHeading = page.getByRole('heading', { name: "You're Live!" });
-    await expect(successHeading).toBeVisible({ timeout: 60000 });
-  });
-
-  // Test 2: Verifies Instant Build handles network error gracefully
-  test('Instant Build gracefully displays an error state on a network failure with proper styling', async ({ page }) => {
-    await page.goto('/onboarding');
-    const setupScreen = page.locator('#setup-screen');
-    await expect(setupScreen).toBeVisible({ timeout: 30000 });
-
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await expect(instantBuildButton).toBeVisible();
-    await instantBuildButton.click();
-
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    await bioInput.fill("Will fail network request");
-
-    await page.route('**/api/onboarding/**', route => route.abort('failed'));
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-    await generateButton.click();
-
-    // Verify error is shown with correct styling
-    const errorBlock = page.locator('.animate-shake').first();
-    await expect(errorBlock).toBeVisible();
-    await expect(errorBlock).toHaveClass(/text-\[#FF3B30\]/);
-    await expect(errorBlock).toHaveClass(/border-\[#FF3B30\]\/30/);
-
-    // Verify textarea has the red border
-    await expect(bioInput).toHaveClass(/border-\[#FF3B30\]/);
-
-    // Typing clears the error border
-    await bioInput.fill("New text");
-    await expect(bioInput).not.toHaveClass(/border-\[#FF3B30\]/);
-
-    await page.unroute('**/api/onboarding/**');
-  });
-
-  // Test 3: Verifies empty input behavior
-  test('Instant Build prevents submission when the input is empty', async ({ page }) => {
-    await page.goto('/onboarding');
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-
-    // Button should be disabled when input is empty.
-    await expect(generateButton).toBeDisabled();
-
-    // We shouldn't see a loading state.
-    const loadingState = page.getByText('Building Your Business...');
-    await expect(loadingState).not.toBeVisible();
-    await expect(page.getByRole('heading', { name: "Tell us about your business" })).toBeVisible();
-  });
-
-  // Test 4: Smart defaults fallback on partial info
-  test('Instant Build handles partial information appropriately by falling back to smart defaults', async ({ page }) => {
-    await page.goto('/onboarding');
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    // Only provide a generic description
-    await bioInput.fill("I sell things online.");
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-    await generateButton.click();
-
-    const successHeading = page.getByRole('heading', { name: "You're Live!" });
-    await expect(successHeading).toBeVisible({ timeout: 60000 });
-  });
-
-  // Test 5: Mobile responsiveness of the Instant Build component
-  test('Instant Build respects mobile viewport constraints (375px) with valid touch targets for the conversational flow', async ({ page }) => {
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/onboarding');
-
-    const instantBuildButton = page.locator('button', { hasText: 'Instant Build' });
-    await instantBuildButton.click();
-
-    const bioInput = page.getByPlaceholder("e.g. I run a local bakery that sells custom vegan cakes...");
-    const box = await bioInput.boundingBox();
-    expect(Math.round(box?.height || 0)).toBeGreaterThanOrEqual(44);
-    expect(box?.width).toBeLessThanOrEqual(375);
-
-    const generateButton = page.getByRole('button', { name: 'Next' });
-    const btnBox = await generateButton.boundingBox();
-    expect(Math.round(btnBox?.height || 0)).toBeGreaterThanOrEqual(44);
-  });
-});
