--- src/ui/next/src/e2e/onboarding.spec.ts
@@ -1,361 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-test.describe('OnboardingWizard CUJ', () => {
-  test.beforeEach(async ({ page }) => {
-    // Clear local storage to ensure fresh state
-    await page.addInitScript(() => {
-      window.localStorage.clear();
-    });
-  });
-
-
-  test('Maya the Baker can complete the onboarding flow', async ({ page }) => {
-
-
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-    await expect(page.getByText("What's the name of your business?")).toBeVisible();
-
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Maya Bakery');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('I bake custom vegan cakes for weddings and parties.');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Portland, OR/i).fill('Seattle, WA');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Everyone');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await expect(page.locator('input[value="I bake custom vegan cakes f..."]')).toBeVisible();
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    await page.getByText('Modern').click();
-    await page.getByPlaceholder(/e.g. Maya Smith/i).fill('Maya Smith');
-    await page.getByPlaceholder(/you@example.com/i).fill('maya@example.com');
-    await page.getByPlaceholder(/••••••••/i).fill('mypassword123');
-
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-    await expect(page.getByText("You're Live!")).toBeVisible();
-    const storedTenantId = await page.evaluate(() => window.localStorage.getItem('tenant_id'));
-    expect(storedTenantId).not.toBeNull();
-  });
-
-  test('Carlos the Handyman sets up his repair business', async ({ page }) => {
-
-
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Carlos Fixes It');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('Plumbing and general repairs');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Portland, OR/i).fill('Austin, TX');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Homeowners');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await expect(page.locator('input[value="Plumbing and general repairs"]')).toBeVisible();
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    await page.getByText('Minimal').click();
-    await page.getByPlaceholder(/e.g. Maya Smith/i).fill('Carlos');
-    await page.getByPlaceholder(/you@example.com/i).fill('carlos@example.com');
-    await page.getByPlaceholder(/••••••••/i).fill('password123');
-
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-    await expect(page.getByText("You're Live!")).toBeVisible();
-    const storedTenantId = await page.evaluate(() => window.localStorage.getItem('tenant_id'));
-    expect(storedTenantId).not.toBeNull();
-  });
-
-  test('Leo the Music Tutor configures online bookings', async ({ page }) => {
-
-
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Leo Guitar Lessons');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('Guitar tutoring online');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Portland, OR/i).fill('Remote');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Students');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await expect(page.locator('input[value="Guitar tutoring online"]')).toBeVisible();
-    // Removed product assertion since fallback logic doesn't generate products
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    await page.getByText('Classic').click();
-    await page.getByPlaceholder(/e.g. Maya Smith/i).fill('Leo Tutor');
-    await page.getByPlaceholder(/you@example.com/i).fill('leo@music.com');
-    await page.getByPlaceholder(/••••••••/i).fill('pass1234');
-
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-    await expect(page.getByText("You're Live!")).toBeVisible();
-    const storedTenantId = await page.evaluate(() => window.localStorage.getItem('tenant_id'));
-    expect(storedTenantId).not.toBeNull();
-  });
-
-  test('Fatima the Food Cart Operator on a slower network', async ({ page }) => {
-
-
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Fatima Halal Food');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('Halal food cart pickup orders');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Portland, OR/i).fill('New York, NY');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Professionals');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await expect(page.locator('input[value="Halal food cart pickup orders"]')).toBeVisible();
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    await page.getByText('Bold').click();
-    await page.getByPlaceholder(/e.g. Maya Smith/i).fill('Fatima');
-    await page.getByPlaceholder(/you@example.com/i).fill('fatima@foodcart.com');
-    await page.getByPlaceholder(/••••••••/i).fill('halal123');
-
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-    await expect(page.getByText("You're Live!")).toBeVisible({ timeout: 5000 });
-    const storedTenantId = await page.evaluate(() => window.localStorage.getItem('tenant_id'));
-    expect(storedTenantId).not.toBeNull();
-  });
-
-  test('User can save a draft and restore it across sessions', async ({ page }) => {
-    let savedWizardState: Record<string, unknown> | undefined;
-
-    // 1. Start Wizard and Save Draft
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-
-    // Check for glassmorphism
-    await expect(page.locator('#setup-screen')).toHaveClass(/glassmorphism/);
-
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('My Restored Business');
-    await page.getByRole('button', { name: 'Save Draft' }).click();
-    await expect(page.getByText('Draft Saved!')).toBeVisible();
-
-    // Ensure localStorage is populated via zustand persist
-    const lsStore = await page.evaluate(() => window.localStorage.getItem('onboarding-storage-v3'));
-    expect(lsStore).toContain('My Restored Business');
-
-    // 2. Clear local storage to simulate device switch
-    await page.evaluate(() => window.localStorage.clear());
-
-    // 3. Reload page and check restoration
-    await page.reload();
-
-    // We should be restored to the first step of the wizard where we were, with the text filled
-    await expect(page.getByText("What's the name of your business?")).toBeVisible();
-    await expect(page.locator('input[value="My Restored Business"]')).toBeVisible();
-  });
-
-  test('Validation errors prevent launching without complete admin info', async ({ page }) => {
-
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Test Business');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('Testing');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Portland, OR/i).fill('Local');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Anyone');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    // Do NOT fill out admin email and password initially
-    await page.getByPlaceholder(/e.g. Maya Smith/i).fill('Test Admin');
-
-    // Attempt to launch store
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-
-    // Expect validation errors to be visible
-    await expect(page.getByText(/is required/i).first()).toBeVisible();
-
-    // Fill in invalid email and password without number
-    await page.getByPlaceholder(/you@example.com/i).fill('invalid-email');
-    await page.getByPlaceholder(/••••••••/i).fill('password');
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-
-    await expect(page.getByText('Please enter a valid email address')).toBeVisible();
-    await expect(page.getByText('Password must be at least 8 characters and contain a number')).toBeVisible();
-
-    // Ensure it hasn't progressed to the success screen
-    await expect(page.getByText("You're Live!")).toBeHidden();
-  });
-
-  test('Submitting empty inputs displays validation errors with visual indicators', async ({ page }) => {
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-
-    // Step 1: Empty Business Name
-    await expect(page.getByText("What's the name of your business?")).toBeVisible();
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('  ');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    const businessNameInput = page.getByPlaceholder(/Maya's Custom Cake/i);
-    await expect(page.getByText('Business Name must be at least 3 characters.')).toBeVisible();
-    await expect(businessNameInput).toHaveClass(/border-\[#FF3B30\]/);
-
-    // Proceed to Step 2
-    await businessNameInput.fill('Valid Business Name');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    // Step 2: Empty What you sell
-    await expect(page.getByText("What do you sell?")).toBeVisible();
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    const whatYouSellInput = page.getByPlaceholder(/I bake custom vegan cakes/i);
-    await expect(page.getByText('Please tell us what you sell.')).toBeVisible();
-    await expect(whatYouSellInput).toHaveClass(/border-\[#FF3B30\]/);
-
-    // Proceed to Step 3
-    await whatYouSellInput.fill('Valid products');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    // Step 3: Empty Location
-    await expect(page.getByText("Where are you located?")).toBeVisible();
-    await page.getByPlaceholder(/Portland, OR/i).fill('  ');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    const locationInput = page.getByPlaceholder(/Portland, OR/i);
-    await expect(page.getByText('Please tell us your location.')).toBeVisible();
-    await expect(locationInput).toHaveClass(/border-\[#FF3B30\]/);
-
-    // Proceed to Step 4
-    await locationInput.fill('Valid Location');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    // Step 4: Empty Target Audience
-    await expect(page.getByText("Who is your target audience?")).toBeVisible();
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('  ');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    const audienceInput = page.getByPlaceholder(/Local families, Tech startups/i);
-    await expect(page.getByText('Please tell us your target audience.')).toBeVisible();
-    await expect(audienceInput).toHaveClass(/border-\[#FF3B30\]/);
-  });
-
-  test('User can use Instant Build to launch storefront quickly', async ({ page }) => {
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-
-    await page.getByRole('button', { name: 'Instant Build' }).click();
-    await expect(page.getByText("Tell us about your business")).toBeVisible();
-
-    const bioInput = page.getByPlaceholder(/e.g. I run a local bakery/i);
-    await bioInput.fill('I am Maya, I run a local bakery making custom vegan cakes in Portland, OR.');
-
-    await page.getByRole('button', { name: 'Generate Storefront' }).click();
-
-    // Expect it to eventually reach "You're Live!" screen
-    await expect(page.getByText("You're Live!")).toBeVisible({ timeout: 15000 });
-  });
-});
-
-  test('Instant Build handles network failures gracefully without mock data', async ({ page, context }) => {
-    await page.goto('/onboarding');
-    await expect(page.getByText("10-Minute Setup Wizard")).toBeVisible();
-    await page.getByRole('button', { name: 'Instant Build' }).click();
-    await expect(page.getByText("Tell us about your business")).toBeVisible();
-
-    // Fill the form
-    await page.getByPlaceholder(/e.g. I run a local bakery/i).fill('Failing business info');
-
-    // Intercept the API route to fail
-    await context.route('/api/onboarding/intake', route => route.abort('failed'));
-
-    await page.getByRole('button', { name: 'Generate Storefront' }).click();
-
-    // Should display a real error message, not mock data
-    await expect(page.getByText(/Failed to launch. Please try again./i)).toBeVisible();
-
-    // Stop interception
-    await context.unroute('/api/onboarding/intake');
-  });
-
-  test('Step-by-step intake handles backend processing errors correctly', async ({ page, context }) => {
-    await page.goto('/onboarding');
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Test Business');
-    await page.getByRole('button', { name: 'Next' }).click();
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('Testing');
-    await page.getByRole('button', { name: 'Next' }).click();
-    await page.getByPlaceholder(/Portland, OR/i).fill('Local');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Testing');
-
-    // Mock the backend responding with a 500 error
-    await context.route('/api/onboarding/intake', route => route.fulfill({ status: 500, json: { error: 'Internal Server Error' } }));
-
-    await page.getByRole('button', { name: 'Next' }).click();
-    await expect(page.getByText(/Internal Server Error/i)).toBeVisible();
-
-    await context.unroute('/api/onboarding/intake');
-  });
-
-  test('Store launch correctly fails when start API is down', async ({ page, context }) => {
-    await page.goto('/onboarding');
-    await page.getByRole('button', { name: 'Start My Business' }).click();
-    await page.getByPlaceholder(/Maya's Custom Cake/i).fill('Test Business');
-    await page.getByRole('button', { name: 'Next' }).click();
-    await page.getByPlaceholder(/I bake custom vegan cakes/i).fill('Testing');
-    await page.getByRole('button', { name: 'Next' }).click();
-    await page.getByPlaceholder(/Portland, OR/i).fill('Local');
-    await page.getByRole('button', { name: 'Next' }).click();
-
-    await page.getByPlaceholder(/Local families, Tech startups/i).fill('Testing');
-
-    // Normal intake response
-    await context.route('/api/onboarding/intake', route => route.fulfill({ status: 200, json: { business_name: 'Test Business', business_type: 'Test', initial_products: [], categories: [] } }));
-    await page.getByRole('button', { name: 'Next' }).click();
-    await expect(page.getByText('Review Details')).toBeVisible();
-    await page.getByRole('button', { name: 'Continue' }).click();
-
-    await page.getByPlaceholder(/e.g. Maya Smith/i).fill('Test Admin');
-    await page.getByPlaceholder(/you@example.com/i).fill('admin@test.com');
-    await page.getByPlaceholder(/••••••••/i).fill('password123');
-
-    // Mock the start API failing
-    await context.route('/api/onboarding/start', route => route.fulfill({ status: 502 }));
-
-    await page.getByRole('button', { name: 'Launch Store' }).click();
-    await expect(page.getByText(/Failed to start onboarding/i)).toBeVisible();
-
-    await context.unroute('/api/onboarding/start');
-    await context.unroute('/api/onboarding/intake');
-  });
