+++ src/ui/tauri/src/ui/setup.html
@@ -1725,7 +1725,7 @@
                   await goToStep('step-approval');
               }
           } catch (e) {
-              chatMessagesEl.innerHTML += `<div style="margin-bottom: 10px; color: #FF3B30;"><strong>Error:</strong> Failed to connect.</div>`;
+              chatMessagesEl.innerHTML += `<div class="animate-shake" style="margin-bottom: 10px; padding: 10px; color: #FF3B30; border: 1px solid rgba(255, 59, 48, 0.3); border-radius: 8px;"><strong>Error:</strong> Failed to connect.</div>`;
           } finally {
               chatSendBtn.disabled = false;
           }
@@ -1931,7 +1931,7 @@
                   const errorMsg = err.message || "Network Error: Failed to fetch";
                   instantErrorEl.innerText = errorMsg;
                   instantErrorEl.style.display = 'block';
-                  instantErrorEl.classList.add('text-[#FF3B30]', 'border-[#FF3B30]/30');
+                  instantErrorEl.classList.add('text-[#FF3B30]', 'border-[#FF3B30]/30', 'animate-shake');
                   instantBioInputEl.classList.add('border-[#FF3B30]');
                   generateStorefrontBtn.disabled = false;
                   generateStorefrontBtn.innerText = "Next";
